requirements.txt
Django==1.11.7
Pillow==4.3.0
django-embed_video==1.1.2
django-user-accounts==1.3.1
django-widget-tweaks==1.4.2
django-mathfilters==0.4.0
selenium==3.11.0
WeasyPrint==0.42.3
cairocffi==0.6
django-filters==0.2.1
django-filter==1.1.0
django-graphos==0.3.41
django-simple-captcha==0.5.9
pytz==2018.5
coverage==4.5.1
django-datetimepicker==3.14
django-easy-pdf==0.1.1